LayoutTests/editing/pasteboard/paste-text-with-style-5.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This tests copying and pasting text does not strip inline styles that overrides UA style rules.
To manually test, copy and paste "hello world" below. The pasted text should not be bolded.</p>
<div id="test" contenteditable><b style="font-weight: normal">hello world</b></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
if (window.testRunner)
     testRunner.dumpAsText();

Markup.description(document.getElementById('description').textContent);

document.getElementById('test').focus();

Markup.dump('test', 'Before copy-paste');
document.execCommand('SelectAll', false, null);
document.execCommand('Copy', false, null);
document.execCommand('Paste', false, null);
Markup.dump('test', 'After copy-paste');

</script>
</body>
</html>
